Refactor na tela de seleção de raça na criação de ficha

+++ api/src/main/java/com/shakal/rpg/api/contracts/entity/IWereable.java
@@ -2,6 +2,6 @@ package com.shakal.rpg.api.contracts.entity;
 
 public interface IWereable extends IBuyable {
 	
-	int weight();
+	double weight();
 
 }
